requirements.txt
python-docx>=0.8.11
markdown>=3.4.0
PyYAML>=6.0